+++ src/app/app.routes.ts
@@ -1,12 +1,12 @@
 import { Routes } from '@angular/router';
 
-import { CenterPageLayoutComponent } from "./common/components/center-page-layout/center-page-layout.component";
-import { LoginComponent } from "./authentication/components/login/login.component";
-import { LogoutComponent } from "./authentication/components/logout/logout.component";
-import { PageNotFoundComponent } from "./page-not-found/components/page-not-found/page-not-found.component";
-import { authGuard } from "./common/guards/authGuard";
-import { DashboardComponent } from "./dashboard/components/dashboard/dashboard.component";
-import { SideNavLayoutComponent } from "./common/components/side-nav-layout/side-nav-layout.component";
+import { CenterPageLayoutComponent } from "./layouts/components/center-page-layout/center-page-layout.component";
+import { LoginComponent } from "./pages/components/login/login.component";
+import { LogoutComponent } from "./pages/components/logout/logout.component";
+import { PageNotFoundComponent } from "./pages/components/page-not-found/page-not-found.component";
+import { authGuard } from "./core/guards/authGuard";
+import { DashboardComponent } from "./pages/components/dashboard/dashboard.component";
+import { SideNavLayoutComponent } from "./layouts/components/side-nav-layout/side-nav-layout.component";
 
 export const routes: Routes = [
   {path: '', redirectTo: 'dashboard', pathMatch: 'full'},
