+++ src/app/layouts/components/side-nav-layout/side-nav-layout.component.scss
@@ -1,4 +1,4 @@
-@import "variables";
+@import "../../../../assets/scss/variables";
 
 main {
   padding-top: $topnav-height + 1.5rem;
