+++ src/app/pages/components/login/login.component.ts
@@ -3,7 +3,7 @@ import { FormsModule } from '@angular/forms';
 import { Router } from "@angular/router";
 import { CommonModule } from "@angular/common";
 
-import { ApiService } from "../../../common/services/api.service";
+import { ApiService } from "../../../core/services/api.service";
 
 @Component({
   selector: 'app-login',
