+++ src/app/shared/components/side-nav/sideNav.component.scss
