+++ src/app/shared/components/top-nav/top-nav.component.scss
